Add basic support for Slack bot. Not yet production ready.
Slack bots can be created using oauth credentials. Botmaster uses
the Slack Events API to receive messages and the Slack web API to send
out messages. Receiving text messages works and sending out text and attachments works (in the same way it would in Messenger).
Not in the same way it would in Slack. I.e. attachments for buttons don't work yet etc... Will get to it. But even in the end,
it will still be a little different. Which is why a _sendSlackFormattedMessage method will be implemented to allow developers
to send a message already formatted for Slack. The same will be done for the other platforms.