Make the BaseBot sendMessage method call three subclasses functions in sequence rather than just __sendMessage
This was done in order to remove __sendRaw from all subclasses as __sendMessage now does that and to make it clear
that before sending a message, an OutgoingMessage object needs to be formatted. After sending a message, a last function
is called: __createStandardBodyResponseComponents in order to add standard body response components to the responseBody that
is ultimately returned. All three methods have to be implemented in all BaseBot subclasses and they all have to return promises
that resolve the expected objects.